% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000-wrappers.R
\name{read_sas}
\alias{read_sas}
\title{Read a SAS7BDAT file into a column-oriented representation}
\usage{
read_sas(path)
}
\arguments{
\item{path}{Path to a .sas7bdat file}
}
\value{
Named list of vectors, with column metadata attached as attributes
}
\description{
Read a SAS7BDAT file into a column-oriented representation
}
