% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000-wrappers.R
\name{sas_column_names}
\alias{sas_column_names}
\title{Column names of a SAS7BDAT file}
\usage{
sas_column_names(path)
}
\arguments{
\item{path}{Path to a .sas7bdat file}
}
\value{
Character vector of column names
}
\description{
Column names of a SAS7BDAT file
}
