% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/000-wrappers.R
\name{sas_metadata_json}
\alias{sas_metadata_json}
\title{Basic metadata as JSON (for convenient consumption in R)}
\usage{
sas_metadata_json(path)
}
\arguments{
\item{path}{Path to a .sas7bdat file}
}
\value{
Length-1 character vector with a JSON string
}
\description{
Basic metadata as JSON (for convenient consumption in R)
}
